backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
xgboost
joblib
pydantic
python-dotenv
apscheduler
pymongo
requests
python-dateutil
pytz
tzdata
dnspython
six
websockets
watchfiles
uvloop
idna
sniffio
anyio
click
h11
httptools
pyyaml
scipy
typing-extensions
typing-inspection
annotated-types
annotated-doc
tzlocal
